Fix benchmark dir
