+++ tritonbench/utils/run_utils.py
@@ -1,12 +1,17 @@
 import copy
 import logging
 import sys
+import time
 import subprocess
 
+from datetime import datetime
+from pathlib import Path
 from tritonbench.utils.env_utils import is_fbcode
 from tritonbench.utils.path_utils import REPO_PATH, remove_cmd_parameter, add_cmd_parameter
 
 from typing import Optional, List
+
+BENCHMARKS_OUTPUT_DIR = REPO_PATH.joinpath(".benchmarks")
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
@@ -31,3 +36,10 @@ def run_in_task(op: str, op_args: Optional[List[str]]=None) -> None:
     except KeyboardInterrupt:
         logger.warning("[tritonbench] KeyboardInterrupt received, exiting...")
         sys.exit(1)
+
+
+def setup_output_dir(bm_name: str):
+    current_timestamp = datetime.fromtimestamp(time.time()).strftime("%Y%m%d%H%M%S")
+    output_dir = BENCHMARKS_OUTPUT_DIR.joinpath(bm_name, f"run-{current_timestamp}")
+    Path.mkdir(output_dir, parents=True, exist_ok=True)
+    return output_dir.absolute()
